requirements.txt
discord.py>=2.3.0
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
ta>=0.11.0
scipy>=1.11.0
matplotlib>=3.7.0
textblob>=0.17.1
python-dateutil>=2.8.0

